% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova-test.R
\name{anova_test}
\alias{anova_test}
\title{(M)ANOVA Permutation Test}
\usage{
anova_test(
  data,
  memberships,
  stats = list(stat_anova_f_ip),
  B = 1000L,
  M = NULL,
  alternative = "right_tail",
  combine_with = "tippett",
  type = "exact",
  seed = NULL,
  ...
)
}
\arguments{
\item{data}{A numeric vector or a numeric matrix or a list specifying the
pooled data points. Alternatively, it can be a distance matrix stored as an
object of class \code{\link[stats:dist]{stats::dist}}, in which case test statistics
based on inter-point distances (marked with the \verb{_ip} suffix) should be
used.}

\item{memberships}{An integer vector specifying the original membership of
each data point.}

\item{stats}{A list of functions produced by \code{\link[rlang:as_function]{rlang::as_function()}}
specifying the chosen test statistic(s). A number of test statistic
functions are implemented in the package and can be used as such.
Alternatively, one can provide its own implementation of test statistics
that (s)he deems relevant for the problem at hand. See the section
\emph{User-supplied statistic function} for more information on how these
user-supplied functions should be structured for compatibility with the
\strong{flipr} framework. Defaults to \code{list(stat_anova_f_ip)}.}

\item{B}{The number of sampled permutations. Default is \code{1000L}.}

\item{M}{The total number of possible permutations. Defaults to \code{NULL}, which
means that it is automatically computed from the given sample size(s).}

\item{alternative}{A single string or a character vector specifying whether
the p-value is right-tailed, left-tailed or two-tailed. Choices are
\code{"right_tail"}, \code{"left_tail"} and \code{"two_tail"}. Default is \code{"two_tail"}. If
a single string is provided, it is assumed that it should be applied to all
test statistics provided by the user. Alternative, the length of
\code{alternative} should match the length of the \code{stats} parameter and it is
assumed that there is a one-to-one correspondence. Defaults to
\code{"right_tail"}.}

\item{combine_with}{A string specifying the combining function to be used to
compute the single test statistic value from the set of p-value estimates
obtained during the non-parametric combination testing procedure. For now,
choices are either \code{"tippett"} or \code{"fisher"}. Defaults to \code{"tippett"},
which picks Tippett's function.}

\item{type}{A string specifying which formula should be used to compute the
p-value. Choices are \code{exact}, \code{upper_bound} and \code{estimate}. See Phipson &
Smith (2010) for details. Defaults to \code{"exact"}.}

\item{seed}{An integer specifying the seed of the random generator useful for
result reproducibility or method comparisons. Default is \code{NULL}.}

\item{...}{Extra parameters specific to some statistics.}
}
\value{
A \code{\link[base:list]{base::list}} with 4 components:
\itemize{
\item \code{observed}: the value of the (possible combined) test statistic(s) using
the original memberhips of data points;
\item \code{pvalue}: the permutation p-value;
\item \code{null_distribution}: the values of the (possible combined) test statistic(s)
using the permuted memberhips of data points;
\item \code{permutations}: the permutations that were effectively sampled to produce
the null distribution.
}
}
\description{
This function carries out an hypothesis test in which the null hypothesis is
that K samples are governed by the same underlying generative probability
distribution against the alternative hypothesis that they are governed by
different generative probability distributions.
}
\section{User-supplied statistic function}{

A user-specified function should have at least two arguments:
\itemize{
\item the first argument should be either a list of the \code{n} pooled data points or
a dissimilarity matrix stored as a \code{\link[stats:dist]{stats::dist}} object.
\item the second argument shoud be an integer vector specifying the (possibly
permuted) membership of each data point.
}

See the \code{\link[=stat_anova_f]{stat_anova_f()}} function for an example.
}

\examples{
out1 <- anova_test(
  data = dist(chickwts$weight),
  memberships = chickwts$feed,
  stats = list(stat_anova_f_ip)
)
out1$pvalue

out2 <- anova_test(
  data = chickwts$weight,
  memberships = chickwts$feed,
  stats = list(stat_anova_f)
)
out2$pvalue
}
