% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFits.R
\name{collisionEnergyLabel}
\alias{collisionEnergyLabel}
\title{Creates the x-axis plot label depending on precursorActivationType and precursorCollisionEnergyUnit as new columns precursorCollisionEnergyUnitLabel.}
\usage{
collisionEnergyLabel(data)
}
\arguments{
\item{data}{the FIP input data.}
}
\description{
Creates the x-axis plot label depending on precursorActivationType and precursorCollisionEnergyUnit as new columns precursorCollisionEnergyUnitLabel.
}
