% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistributions.R
\name{createLipidCreatorParameters}
\alias{createLipidCreatorParameters}
\title{Add parameter for instrumentId (cvTerm)  and output format suitable
for LipidCreator import.}
\usage{
createLipidCreatorParameters(params, instrumentId = "MS:XXXXXX")
}
\arguments{
\item{params}{the fit parameters to modify.}

\item{instrumentId}{the instrumentId cvTerm to use.}
}
\value{
a tibble in the correct format for LipidCreator with ParKey and ParValue.
}
\description{
Add parameter for instrumentId (cvTerm)  and output format suitable
for LipidCreator import.
}
