% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistributions.R
\name{dlnormMode}
\alias{dlnormMode}
\title{#huber loss function
#' @export
huberLossPw <- function(x, hgamma = 0.05) {
  stopifnot(is.double(x))
  if (abs(x) <= hgamma) {
    0.5 * (x) ^ 2
  } else {
    hgamma * abs(x) - ((hgamma ^ 2) / 2)
  }
}}
\usage{
dlnormMode(meanlog, sdlog, shift = 0)
}
\arguments{
\item{meanlog}{the \code{meanlog} argument for \code{dlnorm}.}

\item{sdlog}{the \code{sdlog} argument for \code{dlnorm}}

\item{shift}{the \code{shift} parameter modulates the shift of the log-normal density function or 0.}
}
\value{
the value of the mode function for the given parameters.
}
\description{
#' @export
huberLoss <- function(x, hgamma = 0.05) {
  sapply(x, huberLossPw, hgamma = hgamma)
}
}
\details{
huberLoss2 <- function(x1, x2, hgamma = 0.05) {
  huberLoss(x1 - x2, hgamma = hgamma)
}
Mode function of shifted \code{dlnorm}.
}
