% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistributions.R
\name{dlnormPar}
\alias{dlnormPar}
\title{Parameterized version of the log-normal density function \code{dlnorm} with additional "norm" expression and "scale" factor.
Both are multiplied with the actual log-normal density kernel.}
\usage{
dlnormPar(x, meanlog = 0, sdlog = 1, scale = 1, shift = 0)
}
\arguments{
\item{x}{the \code{x} argument for the log-normal density function}

\item{meanlog}{the \code{meanlog} argument for \code{dlnorm} or 0.}

\item{sdlog}{the \code{sdlog} argument for \code{dlnorm} or 1.}

\item{scale}{the \code{scale} parameter modulates the height of the log-normal density function or 1.}

\item{shift}{the \code{shift} parameter modulates the shift of the log-normal density function on the x-axis or 0.}
}
\value{
The value of the log-normal density function for \code{x} and all parameters.
}
\description{
Parameterized version of the log-normal density function \code{dlnorm} with additional "norm" expression and "scale" factor.
Both are multiplied with the actual log-normal density kernel.
}
