% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistributions.R
\name{fits}
\alias{fits}
\title{Calculates the nonlinear fits for the given tibble. Data will be grouped by
combinationId, species, fragment, adduct, polarity, calculatedMass, foundMassRange[ppm], and group columns.
Calculates non linear regression models by performing an interative grid search within the coordinate bounds provided
by \code{lower} and \code{upper} vectors, starting at \code{start_lower} and \code{start_upper}.
Intermediate models are scored by the AIC value until convergence has been achieved as to the defaults of nls.multstart.}
\usage{
fits(tibble, outputPrefix, group, instrumentId, skipGroupOutput = TRUE,
  start_lower, start_upper, lower, upper,
  minSamplesPerCombinationId = 50)
}
\arguments{
\item{tibble}{the data tibble containing fragment data.}

\item{outputPrefix}{the output prefix for the file.}

\item{group}{the group that this data belongs to.}

\item{instrumentId}{the instrumentId of this data.}

\item{skipGroupOutput}{whether to skip writing of grouping information (for debugging).}

\item{start_lower}{the lower bound to start the parameter grid search.}

\item{start_upper}{the upper bound to start the parameter grid search.}

\item{lower}{the lower bound for the parameter grid search.}

\item{upper}{the upper bound for the parameter grid search.}

\item{minSamplesPerCombinationId}{the minimum number of data points required per fragment / adduct / ppm combination to be considered for model calculation.}
}
\description{
Calculates the nonlinear fits for the given tibble. Data will be grouped by
combinationId, species, fragment, adduct, polarity, calculatedMass, foundMassRange[ppm], and group columns.
Calculates non linear regression models by performing an interative grid search within the coordinate bounds provided
by \code{lower} and \code{upper} vectors, starting at \code{start_lower} and \code{start_upper}.
Intermediate models are scored by the AIC value until convergence has been achieved as to the defaults of nls.multstart.
}
