% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\title{Flipping data frames.}
\usage{
flip(df, tib)
}
\arguments{
\item{df}{Short for data frame, the first argument should be a data frame of any type.}

\item{tib}{Short for tibble, a logical argument describing if data should be returned
in tibble format or not. Default is FALSE.}
}
\value{
Either a data frame or tibble with columns and rows flipped.
}
\description{
This function flips rows and columns in a data frame and optionally returns
the flipped data frame as a tibble.
}
\details{
Function will return errors if first argument is not of type data frame. tib is
an optional argument that will default to FALSE if left empty.
}
