% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipr-package.R
\docType{package}
\name{flipr-package}
\alias{flipr}
\alias{flipr-package}
\title{flipr: Flexible inference via permutations in R}
\description{
The flipr package provides a flexible permutation framework for making
inference such as point estimation, confidence intervals or hypothesis
testing, on any kind of data, be it univariate, multivariate, or more complex
such as network-valued data, topological data, functional data or
density-valued data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://LMJL-Alea.github.io/flipr/}
  \item \url{https://github.com/LMJL-Alea/flipr/}
  \item Report bugs at \url{https://github.com/LMJL-Alea/flipr/issues/}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})

Authors:
\itemize{
  \item Alessia Pini \email{alessia.pini@unicatt.it}
  \item Simone Vantini \email{simone.vantini@polimi.it}
}

Other contributors:
\itemize{
  \item Juliette Chiapello [contributor]
}

}
\keyword{internal}
