% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipr-package.R
\docType{package}
\name{flipr}
\alias{flipr}
\title{flipr: Flexible inference via permutations in R}
\description{
The flipr package provides a flexible permutation framework for making
inference such as point estimation, confidence intervals or hypothesis
testing, on any kind of data, be it univariate, multivariate, or more complex
such as network-valued data, topological data, functional data or
density-valued data.
}
