% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ranges}
\alias{get_ranges}
\title{Extract ranges from parameter list}
\usage{
get_ranges(param_list)
}
\arguments{
\item{param_list}{An object of class \code{\link[dials]{parameters}} storing
information about the parameters under investigation.}
}
\value{
A list of length-2 numeric vectors specifying the ranges for each
parameter under investigation.
}
\description{
This function extracts the ranges of a parameter list created via
\code{\link[dials]{parameters}} into a list of length-2 numeric vectors.
}
\examples{
x <- rnorm(10)
y <- rnorm(10, mean = 2)
null_spec <- function(y, parameters) {
  purrr::map(y, ~ .x - parameters[1])
}
stat_functions <- list(stat_t)
stat_assignments <- list(mean_param = 1)
pf <- PlausibilityFunction$new(
  null_spec = null_spec,
  stat_functions = stat_functions,
  stat_assignments = stat_assignments,
  x, y
)
pf$set_point_estimates(point_estimate = mean(y) - mean(x))
pf$set_parameter_bounds(
  point_estimate = pf$point_estimates,
  conf_level = pf$max_conf_level
)
get_ranges(pf$param_list)
}
