% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_row.R
\name{invert_row}
\alias{invert_row}
\title{Inverting data frame rows.}
\usage{
invert_row(df, out)
}
\arguments{
\item{df}{Short for data frame, the first argument should be a data frame of any type.}

\item{out}{Short for output, this argument determines the form of the resulting data.
Options include 'DF' for data frame, 'TB' for tibble, and 'DT' for data table.}
}
\value{
Either a data frame, data table, or tibble with row order inverted.
}
\description{
This function will invert the order of rows in a data frame and optionally
returns the data frame as a tibble or data table.
}
\details{
Function will return errors if first argument is not of type data frame. out is
an optional argument that determines if the output is a data frame, tibble, or data table.
Defaults to data frame if left empty.
}
\examples{
invert_row(mtcars)
invert_row(mtcars, "TB")
}
