% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_pf}
\alias{plot_pf}
\title{Visualization of Plausibility Functions}
\usage{
plot_pf(pf, ngrid = 100, ncores = 1)
}
\arguments{
\item{pf}{A \code{\link{PlausibilityFunction}} object.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
This function plots the plausibility function for up to two parameters of
interest.
}
\examples{
x <- rnorm(10)
y <- rnorm(10, mean = 2)
null_spec <- function(y, parameters) {purrr::map(y, ~ .x - parameters[1])}
stat_functions <- list(stat_t)
stat_assignments <- list(mean_param = 1)
pf <- PlausibilityFunction$new(
  null_spec = null_spec,
  stat_functions = stat_functions,
  stat_assignments = stat_assignments,
  x, y
)
plot_pf(pf)
}
