% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFits.R
\name{plotParameterConfidenceIntervals}
\alias{plotParameterConfidenceIntervals}
\title{Plots the confidence intervals for the parameters used for model optimization.}
\usage{
plotParameterConfidenceIntervals(nlsFitPlotsOutputList, combinationId,
  outputPrefix, plotFormat = "png", plotDimensions = list(width =
  11.69, height = 8.27), color_scale = ggplot2::scale_colour_hue())
}
\arguments{
\item{nlsFitPlotsOutputList}{the FIP fits output list.}

\item{combinationId}{the combinationId for the case to plot.}

\item{outputPrefix}{the basename (identifying a case) for the plot.}

\item{plotFormat}{the format, passed to \code{ggplot2::ggsave}.}

\item{plotDimensions}{the dimensions of the plot.}

\item{color_scale}{the shared color scale to identify fragment and adduct pairs.}
}
\description{
Plots the confidence intervals for the parameters used for model optimization.
}
