% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotPrecCollEnergyVsScanRelativeIntensityOverlay}
\alias{plotPrecCollEnergyVsScanRelativeIntensityOverlay}
\title{Plots precursor collision energy vs. total scan intensity normalized relative fragment intensity overlaid on one panel.}
\usage{
plotPrecCollEnergyVsScanRelativeIntensityOverlay(data, basename,
  plotFormat = "png", plotDimensions = list(width = 11.69, height =
  8.27), color_scale = ggplot2::scale_colour_hue())
}
\arguments{
\item{data}{the FIP data to plot.}

\item{basename}{the basename (identifying a case) for the plot.}

\item{plotFormat}{the format, passed to \code{ggplot2::ggsave}.}

\item{plotDimensions}{the dimensions of the plot.}

\item{color_scale}{the shared color scale to identify fragment and adduct pairs.}
}
\value{
the ggplot object.
}
\description{
Plots precursor collision energy vs. total scan intensity normalized relative fragment intensity overlaid on one panel.
}
