% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFits.R
\name{plotResiduals}
\alias{plotResiduals}
\title{Plots the residuals between predicted and measured values.}
\usage{
plotResiduals(nlsFitPlotsOutputList, outputPrefix, plotFormat = "png",
  plotDimensions = list(width = 11.69, height = 8.27),
  color_scale = ggplot2::scale_colour_hue())
}
\arguments{
\item{nlsFitPlotsOutputList}{the FIP fits output list.}

\item{outputPrefix}{the basename (identifying a case) for the plot.}

\item{plotFormat}{the format, passed to \code{ggplot2::ggsave}.}

\item{plotDimensions}{the dimensions of the plot.}

\item{color_scale}{the shared color scale to identify fragment and adduct pairs.}
}
\description{
Plots the residuals between predicted and measured values.
}
