% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation_data}
\alias{simulation_data}
\title{Test statistics of a phylogenetic model test.}
\format{
A data frame with 6 rows and 5 variables:
\describe{
\item{1}{First simulation, unitless}
\item{2}{Second simulation, unitless}
\item{3}{Third simulation, unitless}
\item{4}{Fourth simulation, unitless}
\item{5}{Fifth simulation, unitless}

}
}
\source{
Simulated using R package arbutus
}
\usage{
simulation_data
}
\description{
A dataset containing the 5 simulations of the p values for a set of test statistics
generated to determine the adequacy of a phylogenetic model for a given set of
character data.
}
\keyword{datasets}
