% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-sample-ci.R
\name{two_sample_ci}
\alias{two_sample_ci}
\title{Two-Sample Permutation Confidence Interval}
\usage{
two_sample_ci(
  point_estimate,
  alpha,
  null_specification,
  x,
  y,
  statistic = stat_hotelling,
  B = 1000L,
  type = "exact",
  combine_with = "tippett",
  seed = NULL,
  alternative = "right_tail"
)
}
\arguments{
\item{point_estimate}{A scalar value providing a point estimate of the
parameter for which we want to compute a confidence interval.}

\item{alpha}{A scalar specifying the desired significance level.}

\item{null_specification}{A function with two arguments \code{y} and \code{parameters}
such that \verb{F_X = F_\{null_specification(Y, parameters)\}} under the null
hypothesis.}

\item{x}{A list or matrix representing the 1st sample.}

\item{y}{A list or matrix representing the 2nd sample.}

\item{statistic}{A character vector specifying the chosen test statistic(s).
These can be \code{\link{stat_hotelling}} or user-specified functions that
define desired statistics. See the section \emph{User-supplied statistic
function} for more information on how these user-supplied functions should
be structured for compatibility with the \strong{flipr} framwork. Default is
\code{\link{stat_hotelling}}.}

\item{B}{The number of sampled permutation. Default is \code{1000L}.}

\item{type}{A string specifying if performing an exact test through the use
of Phipson-Smyth estimate of the p-value or an approximate test through a
Monte-Carlo estimate of the p-value. Default is \code{"exact"}.}

\item{combine_with}{A string specifying the combining function to be used to
compute the single test statistic value from the set of p-value estimates
obtained during the non-parametric combination testing procedure. Default
is \code{"tippett"}, which picks Tippett's function.}

\item{seed}{An integer specifying the seed of the random generator useful for
result reproducibility or method comparisons. Default is \code{NULL}.}

\item{alternative}{A string specifying whether the p-value is right-tailed,
left-tailed or two-tailed. Choices are \code{"right_tail"}, \code{"left_tail"} and
\code{"two_tail"}. Default is \code{"right_tail"}. Obviously, if the test statistic
used in argument \code{statistic} is positive, all alternatives will lead to the
two-tailed p-value.}
}
\value{
A length-2 vector providing the confidence interval at the required
level.
}
\description{
This function calculates permutation confidence intervals for two-sample
problems. This is done through the specification a point estimate of the
parameter to be estimated and of a set of null hypotheses of the form \verb{F_X = F_\{g(Y, parameters)\}} where \code{g} is a user-supplied function.
}
\examples{
x1 <- rnorm(10)
x2 <- rnorm(10, 3)
null_spec <- function(y, parameters) {y - parameters[1]}
two_sample_ci(
  point_estimate = mean(x2) - mean(x1),
  alpha = 0.05,
  null_specification = null_spec,
  x = x1,
  y = x2,
  statistic = stat_t,
  seed = 1234,
  B = 1000,
  alternative = "two_tail"
)
}
