% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoSample.R
\name{two_sample_test}
\alias{two_sample_test}
\title{Two-Sample Permutation Test}
\usage{
two_sample_test(x, y, statistic = "hotelling", B = 1000L,
  alpha = 0.05, test = "exact", seed = NULL)
}
\arguments{
\item{x}{A list or matrix representing the 1st sample.}

\item{y}{A list or matrix representing the 2nd sample.}

\item{statistic}{A string specifying the chosen test statistic(s), among:
\code{"hotelling"} [default].}

\item{B}{The number of sampled permutation (default: \code{1000L}).}

\item{alpha}{The significance level (default: \code{0.05}).}

\item{test}{A character string specifying if performing an exact test through
the use of Phipson-Smyth estimate of the p-value or an approximate test
through a Monte-Carlo estimate of the p-value (default: \code{"exact"}).}

\item{seed}{An integer for specifying the seed of the random generator for
result reproducibility or method comparisons (default: \code{NULL}).}
}
\value{
A \code{\link[base]{list}} with three components: the value of the
  statistic for the original two samples, the p-value of the resulting
  permutation test and a numeric vector storing the values of the permuted
  statistics.
}
\description{
This function carries out an hypothesis test where the null hypothesis is
that the two samples are ruled by the same underyling generative probability
distributions against the alternative hypothesis that they are ruled by two
separate generative probability distributions.
}
\examples{
n <- 10L
p <- 3L
mx <- rep(0, p)
sigma <- diag(1, p)

# Two different models for the two populations
x <- mvnfast::rmvn(n = n, mu = mx, sigma = sigma)
delta <- 10
my <- mx + delta
y <- mvnfast::rmvn(n = n, mu = my, sigma = sigma)
t1 <- two_sample_test(x, y)
t1$pvalue

# Same model for the two populations
x <- mvnfast::rmvn(n = n, mu = mx, sigma = sigma)
delta <- 0
my <- mx + delta
y <- mvnfast::rmvn(n = n, mu = my, sigma = sigma)
t2 <- two_sample_test(x, y)
t2$pvalue
}
