% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-stats.R
\name{use_stat}
\alias{use_stat}
\title{Test Statistic Template}
\usage{
use_stat(nsamples = 1, stat_name = "mystat")
}
\arguments{
\item{nsamples}{An integer specifying the number of samples to be used.
Defaults to \code{1L}. Currently only works for one- or two-sample problems.}

\item{stat_name}{A string specifying the name of the test statistic that is
being implemented. Defaults to \code{mystat}.}
}
\value{
Creates a dedicated \code{.R} file with a template of code for the
function that implements the test statistic and saves it to the \verb{R/} folder
of your package.
}
\description{
This function is a helper to automatically generate an \code{.R} file populated
with a skeleton of a typical test function compatible with \code{flipr}.
}
\examples{
\dontrun{
use_stat()
}
}
