% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDistributions.R
\name{writeGroup}
\alias{writeGroup}
\title{Selects \code{combinationId}, \code{precursorCollisionEnergy} and \code{scanRelativeIntenty} to write to file.}
\usage{
writeGroup(tibble, outputPrefix, skip = FALSE)
}
\arguments{
\item{tibble}{the data frame / tibble to write.}

\item{outputPrefix}{the output prefix for the file.}

\item{skip}{if true, skip writing.}
}
\value{
the input tibble.
}
\description{
Selects \code{combinationId}, \code{precursorCollisionEnergy} and \code{scanRelativeIntenty} to write to file.
}
