% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-stat.R
\name{regression-stats}
\alias{regression-stats}
\alias{stat_regression_global}
\alias{stat_regression_coef}
\title{Test Statistics for the Linear Regression Problem}
\usage{
stat_regression_global(vars, response, ...)

stat_regression_coef(vars, response, coef, ...)
}
\arguments{
\item{vars}{A data frame of all variables used as predictors in the model.}

\item{response}{An numeric vector representing the response variable.}

\item{...}{Extra parameters specific to some statistics.}

\item{coef}{For tests on one coefficient, an integer specifying the index of
the coefficient to be tested.}
}
\value{
A numeric value storing the value of test statistic given the
(possibly permuted) response specified by \code{response}.
}
\description{
This is a collection of functions that provide test statistics to be used
into the permutation scheme for performing Linear Regression. These test
statistics can be divided into two categories: traditional statistics that
use empirical moments and inter-point statistics that only rely on pairwise
dissimilarities between data points.
}
\section{Traditional Test Statistics}{

\itemize{
\item \code{\link[=stat_regression_global]{stat_regression_global()}} implements the F statistic used to test
globally the coefficients of the regression.
\item \code{\link[=stat_regression_coef]{stat_regression_coef()}} implements the t statistic used to test one
coefficient of the regression.
}
}

\examples{
response_var <- iris$Sepal.Length
predictors <- iris[2:5]
stat_regression_global(predictors, response_var)
stat_regression_coef(predictors, response_var, coef = 1)

}
