% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-test.R
\name{regression_test}
\alias{regression_test}
\title{Linar Regression Permutation Test}
\usage{
regression_test(
  original_response,
  fitted,
  residuals,
  vars,
  index,
  stats = list(stat_lm),
  B = 1000L,
  M = NULL,
  alternative = "right_tail",
  combine_with = "tippett",
  type = "exact",
  seed = NULL,
  ...
)
}
\arguments{
\item{original_response}{An numeric vector representing the response variable.}

\item{fitted}{The fitted values of the reduced model.}

\item{residuals}{The residual values of the reduced model.}

\item{vars}{A data frame of all variables used as predictors in the model.}

\item{index}{For tests on one coefficient, an integer specifying the index of
the coefficient to be tested.}

\item{stats}{A list of functions produced by \code{\link[rlang:as_function]{rlang::as_function()}}
specifying the chosen test statistic(s). A number of test statistic
functions are implemented in the package and can be used as such.
Alternatively, one can provide its own implementation of test statistics
that (s)he deems relevant for the problem at hand. See the section
\emph{User-supplied statistic function} for more information on how these
user-supplied functions should be structured for compatibility with the
\strong{flipr} framework. Defaults to \code{list(stat_lm)}.}

\item{B}{The number of sampled permutations. Default is \code{1000L}.}

\item{M}{The total number of possible permutations. Defaults to \code{NULL}, which
means that it is automatically computed from the given sample size(s).}

\item{alternative}{A single string or a character vector specifying whether
the p-value is right-tailed, left-tailed or two-tailed. Choices are
\code{"right_tail"}, \code{"left_tail"} and \code{"two_tail"}. Default is \code{"two_tail"}. If
a single string is provided, it is assumed that it should be applied to all
test statistics provided by the user. Alternative, the length of
\code{alternative} should match the length of the \code{stats} parameter and it is
assumed that there is a one-to-one correspondence. Defaults to
\code{"right_tail"}.}

\item{combine_with}{A string specifying the combining function to be used to
compute the single test statistic value from the set of p-value estimates
obtained during the non-parametric combination testing procedure. For now,
choices are either \code{"tippett"} or \code{"fisher"}. Defaults to \code{"tippett"},
which picks Tippett's function.}

\item{type}{A string specifying which formula should be used to compute the
p-value. Choices are \code{exact}, \code{upper_bound} and \code{estimate}. See Phipson &
Smith (2010) for details. Defaults to \code{"exact"}.}

\item{seed}{An integer specifying the seed of the random generator useful for
result reproducibility or method comparisons. Default is \code{NULL}.}

\item{...}{Extra parameters specific to some statistics.}
}
\value{
A \code{\link[base:list]{base::list}} with 4 components:
\itemize{
\item \code{observed}: the value of the (possible combined) test statistic(s) using
the original memberhips of data points;
\item \code{pvalue}: the permutation p-value;
\item \code{null_distribution}: the values of the (possible combined) test statistic(s)
using the permuted memberhips of data points;
\item \code{permutations}: the permutations that were effectively sampled to produce
the null distribution.
}
}
\description{
This function carries out an hypothesis test in which the null hypothesis is
that predictors don't have an effect on the response variable against the
alternative hypothesis that they do have an effect on the response variable.
}
\section{User-supplied statistic function}{

A user-specified function should have at least two arguments:
\itemize{
\item the first argument should be a numeric vector representing the response
variable.
\item the second argument should be a data drame of the \code{n} pooled data points.
\item the third argument is optional and represent the index of the coefficient
to be tested in the case of a test on a specific coefficient.
}

See the \code{\link[=stat_lm]{stat_lm()}} function for an example.
}

\examples{
Y <- iris$Sepal.Length
regressors <- iris[2:5]
reduced_vars <- regressors[- c(1)]
fit <- lm(Y ~ ., data = reduced_vars)
out2 <- regression_test(
  original_response = Y,
  fitted <- fit$fitted,
  residuals <- fit$residuals,
  vars = iris[2:5],
  index = 1,
  stats = list(stat_lm)
)
out2$pvalue
}
